Refactoring: suddivisione in moduli e aggiornamento README
