+++ .gitignore
@@ -1,0 +1,3 @@
+venv/
+__pycache__/
+*.pyc
