+++ main.py
@@ -1,0 +1,7 @@
+from tkinter import Tk
+from gui import WhatsAppForensicsApp
+
+if __name__ == "__main__":
+    root = Tk()
+    app = WhatsAppForensicsApp(root)
+    root.mainloop()
